pyproject.toml
[project]
name = "xiaohongshu-generator"
version = "0.1.0"
description = "小红书图文生成器"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "python-dotenv>=1.0.0",
    "google-genai>=1.0.0",
    "pyyaml>=6.0.0",
    "requests>=2.31.0",
    "pillow>=12.0.0",
    # 异步任务队列支持
    "redis>=5.0.0",
    "rq>=1.16.0",
    # 数据库和认证支持
    "sqlalchemy>=2.0.0",
    "bcrypt>=4.0.0",
    "pyjwt>=2.8.0",
    "pymysql>=1.1.0",  # MySQL 支持（可选）
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[tool.uv]
dev-dependencies = []
